Rename call-super-in-init rule to require-super-in-init
