+++ lib/recommended-rules.js
@@ -6,7 +6,7 @@
 module.exports = {
   "ember/alias-model-in-controller": "error",
   "ember/avoid-leaking-state-in-components": "error",
-  "ember/call-super-in-init": "error",
+  "ember/require-super-in-init": "error",
   "ember/closure-actions": "error",
   "ember/jquery-ember-run": "error",
   "ember/local-modules": "off",
