+++ lib/rules/require-super-in-init.js
